Refactoring for layout components
