+++ components/layout-components/Navbar.tsx
@@ -1,58 +1,20 @@
 import React, { ReactNode } from 'react';
 import {
-	IconButton,
-	Avatar,
 	Box,
-	CloseButton,
-	Flex,
-	HStack,
-	VStack,
-	Icon,
 	useColorModeValue,
-	Link,
 	Drawer,
 	DrawerContent,
-	Text,
 	useDisclosure,
-	BoxProps,
-	FlexProps,
-	Menu,
-	MenuButton,
-	MenuDivider,
-	MenuItem,
-	MenuList,
 } from '@chakra-ui/react';
-import {
-	FiHome,
-	FiTrendingUp,
-	FiCompass,
-	FiStar,
-	FiSettings,
-	FiMenu,
-	FiBell,
-	FiChevronDown,
-} from 'react-icons/fi';
-import { IconType } from 'react-icons';
-import { ReactText } from 'react';
-import { AuthContext } from '../../context/authContext';
+import SidebarContent from './SidebarContent';
+import MobileNav from './MobileNav';
 
-interface LinkItemProps {
-	name: string;
-	icon: IconType;
-}
-const LinkItems: Array<LinkItemProps> = [
-	{ name: 'Home', icon: FiHome },
-	{ name: 'Trending', icon: FiTrendingUp },
-	{ name: 'Explore', icon: FiCompass },
-	{ name: 'Favourites', icon: FiStar },
-	{ name: 'Settings', icon: FiSettings },
-];
 
-export default function SidebarWithHeader({
+const SidebarWithHeader = ({
 	children,
 }: {
 	children: ReactNode;
-}) {
+}) => {
 	const { isOpen, onOpen, onClose } = useDisclosure();
 	return (
 		<Box minH="100vh" bg={useColorModeValue('gray.100', 'gray.900')}>
@@ -81,151 +43,4 @@ export default function SidebarWithHeader({
 	);
 }
 
-interface SidebarProps extends BoxProps {
-	onClose: () => void;
-}
-
-const SidebarContent = ({ onClose, ...rest }: SidebarProps) => {
-	return (
-		<Box
-			transition="3s ease"
-			bg={useColorModeValue('white', 'gray.900')}
-			borderRight="1px"
-			borderRightColor={useColorModeValue('gray.200', 'gray.700')}
-			w={{ base: 'full', md: 60 }}
-			pos="fixed"
-			h="full"
-			{...rest}>
-			<Flex h="20" alignItems="center" mx="8" justifyContent="space-between">
-				<Text fontSize="2xl" fontFamily="monospace" fontWeight="bold">
-					MyHotel
-				</Text>
-				<CloseButton display={{ base: 'flex', md: 'none' }} onClick={onClose} />
-			</Flex>
-			{LinkItems.map((link) => (
-				<NavItem key={link.name} icon={link.icon}>
-					{link.name}
-				</NavItem>
-			))}
-		</Box>
-	);
-};
-
-interface NavItemProps extends FlexProps {
-	icon: IconType;
-	children: ReactText;
-}
-const NavItem = ({ icon, children, ...rest }: NavItemProps) => {
-	return (
-		<Link href="#" style={{ textDecoration: 'none' }} _focus={{ boxShadow: 'none' }}>
-			<Flex
-				align="center"
-				p="4"
-				mx="4"
-				borderRadius="lg"
-				role="group"
-				cursor="pointer"
-				_hover={{
-					bg: 'cyan.400',
-					color: 'white',
-				}}
-				{...rest}>
-				{icon && (
-					<Icon
-						mr="4"
-						fontSize="16"
-						_groupHover={{
-							color: 'white',
-						}}
-						as={icon}
-					/>
-				)}
-				{children}
-			</Flex>
-		</Link>
-	);
-};
-
-interface MobileProps extends FlexProps {
-	onOpen: () => void;
-}
-const MobileNav = ({ onOpen, ...rest }: MobileProps) => {
-	const authContext = React.useContext(AuthContext);
-
-	return (
-		<Flex
-			ml={{ base: 0, md: 60 }}
-			px={{ base: 4, md: 4 }}
-			height="20"
-			alignItems="center"
-			bg={useColorModeValue('white', 'gray.900')}
-			borderBottomWidth="1px"
-			borderBottomColor={useColorModeValue('gray.200', 'gray.700')}
-			justifyContent={{ base: 'space-between', md: 'flex-end' }}
-			{...rest}>
-			<IconButton
-				display={{ base: 'flex', md: 'none' }}
-				onClick={onOpen}
-				variant="outline"
-				aria-label="open menu"
-				icon={<FiMenu />}
-			/>
-
-			<Text
-				display={{ base: 'flex', md: 'none' }}
-				fontSize="2xl"
-				fontFamily="monospace"
-				fontWeight="bold">
-				MyHotel
-			</Text>
-
-			<HStack spacing={{ base: '0', md: '6' }}>
-				<IconButton
-					size="lg"
-					variant="ghost"
-					aria-label="open menu"
-					icon={<FiBell />}
-				/>
-				<Flex alignItems={'center'}>
-					<Menu>
-						<MenuButton
-							py={2}
-							transition="all 0.3s"
-							_focus={{ boxShadow: 'none' }}>
-							<HStack>
-								<Avatar
-									size={'sm'}
-									src={
-										'https://images.unsplash.com/photo-1619946794135-5bc917a27793?ixlib=rb-0.3.5&q=80&fm=jpg&crop=faces&fit=crop&h=200&w=200&s=b616b2c5b373a80ffc9636ba24f7a4a9'
-									}
-								/>
-								<VStack
-									display={{ base: 'none', md: 'flex' }}
-									alignItems="flex-start"
-									spacing="1px"
-									ml="2">
-									<Text fontSize="sm">{authContext.user?.displayName ?? 'Unknown User'}</Text>
-									<Text fontSize="xs" color="gray.600">
-										Admin
-									</Text>
-								</VStack>
-								<Box display={{ base: 'none', md: 'flex' }}>
-									<FiChevronDown />
-								</Box>
-							</HStack>
-						</MenuButton>
-						<MenuList
-							bg={useColorModeValue('white', 'gray.900')}
-							borderColor={useColorModeValue('gray.200', 'gray.700')}>
-							<MenuItem>Profile</MenuItem>
-							<MenuItem>Settings</MenuItem>
-							<MenuItem>Billing</MenuItem>
-							<MenuDivider />
-							<MenuItem onClick={authContext.logout}>Sign out</MenuItem>
-						</MenuList>
-					</Menu>
-				</Flex>
-			</HStack>
-		</Flex>
-	);
-};
+export default SidebarWithHeader;
