+++ components/layout-components/SidebarContent.tsx
@@ -1,0 +1,36 @@
+import { Flex, useColorModeValue, Text, CloseButton, Box, BoxProps } from '@chakra-ui/react';
+import { SIDEBAR_LINKS } from '../../constants';
+import NavItem from './NavItem';
+import { APP_NAME } from '../../constants/index';
+
+interface SidebarProps extends BoxProps {
+	onClose: () => void;
+}
+
+const SidebarContent = ({ onClose, ...rest }: SidebarProps) => {
+	return (
+		<Box
+			transition="3s ease"
+			bg={useColorModeValue('white', 'gray.900')}
+			borderRight="1px"
+			borderRightColor={useColorModeValue('gray.200', 'gray.700')}
+			w={{ base: 'full', md: 60 }}
+			pos="fixed"
+			h="full"
+			{...rest}>
+			<Flex h="20" alignItems="center" mx="8" justifyContent="space-between">
+				<Text fontSize="2xl" fontFamily="monospace" fontWeight="bold">
+					{APP_NAME}
+				</Text>
+				<CloseButton display={{ base: 'flex', md: 'none' }} onClick={onClose} />
+			</Flex>
+			{SIDEBAR_LINKS.map((link) => (
+				<NavItem key={link.name} icon={link.icon}>
+					{link.name}
+				</NavItem>
+			))}
+		</Box>
+	);
+};
+
+export default SidebarContent;
